Add morse example using user defined effect
