+++ platformio.ini
@@ -9,12 +9,14 @@
 ; http://docs.platformio.org/page/projectconf.html
 
 [platformio]
-;env_default = nanoatmega328
-env_default = nucleo_f401re
+env_default = nanoatmega328
+;env_default = nucleo_f401re
 ;env_default = esp8266;
 ;env_default = esp32
+
 ; uncomment example to build
 src_dir = examples/hello
+;src_dir = examples/morse
 ;src_dir = examples/breathe
 ;src_dir = examples/fade_on
 ;src_dir = examples/fade_off
